Conflicts:
	Core/WCell.Core/Database/WCellRecord.cs
	Services/WCell.RealmServer/Achievements/AchievementProgressRecord.cs
	Services/WCell.RealmServer/Achievements/AchievementRecord.cs
	Services/WCell.RealmServer/Database/AccountDataRecord.cs
	Services/WCell.RealmServer/Database/EquipmentSet.cs
	Services/WCell.RealmServer/Database/EquipmentSetItemMapping.cs
	Services/WCell.RealmServer/Database/Guild.cs
	Services/WCell.RealmServer/Database/GuildBankTab.cs
	Services/WCell.RealmServer/Database/GuildBankTabItemMapping.cs
	Services/WCell.RealmServer/Database/GuildEventLogEntry.cs
	Services/WCell.RealmServer/Database/ReputationRecord.cs
	Utilities/WCell.Util/Threading/TaskParallel/BlockingCollectionExtensions.cs

+++ Services/WCell.RealmServer/Entities/NamedEntity.cs
@@ -1,3 +1,9 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using WCell.RealmServer.Database.Entities;
+using WCell.Util;
 using WCell.Core;
 using WCell.RealmServer.Database;
 using WCell.Util;
@@ -11,7 +17,7 @@ namespace WCell.RealmServer.Entities
 	{
 		public static EntityId CreateId()
 		{
-			return EntityId.GetPlayerId(CharacterRecord.NextId());
+			return EntityId.GetPlayerId((uint)CharacterRecord.NextId());
 		}
